Fixed Employee Takeout
